+++ smart-sso/smart-sso-demo/src/main/webapp/WEB-INF/views/index.jsp
