+++ smart-sso/smart-sso-demo/src/main/webapp/WEB-INF/web.xml
@@ -34,13 +34,18 @@
 		<filter-name>smartFilter</filter-name>
 		<url-pattern>/*</url-pattern>
 	</filter-mapping>
-	<!-- 首页 -->
+	<!-- SpringMVC核心 -->
 	<servlet>
-		<servlet-name>AdminServlet</servlet-name>
-		<servlet-class>com.smart.sso.demo.AdminServlet</servlet-class>
+		<servlet-name>dispatcher</servlet-name>
+		<servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
+		<init-param>
+			<param-name>contextConfigLocation</param-name>
+			<param-value>classpath:applicationContext-mvc.xml</param-value>
+		</init-param>
+		<load-on-startup>1</load-on-startup>
 	</servlet>
 	<servlet-mapping>
-		<servlet-name>AdminServlet</servlet-name>
+		<servlet-name>dispatcher</servlet-name>
 		<url-pattern>/</url-pattern>
 	</servlet-mapping>
 </web-app>
